==> code/02_standardise_footprints.ipynb <==
# Standardise building footprints

Ensure that all the building footprints are stored in a standardised form.

import geopandas as gpd
import pandas as pd

DATADIR = "/data/uscuni-ulce/buildings_standardised/"

regions = [
    "buildings_austria.pq",
    "buildings_slovakia.pq",
    "buildings_poland.pq",
    "buildings_czechia.pq",
]

### Germany

ger_regions = [
    "bavaria",
    "brandeburg",
    "bw",
    "hessen",
    "mv",
    "nrw",
    "rlp",
    "saarland",
    "sachsen",
    "sachsen-anhalt",
    "sh",
    "th",
    "ni",
    "bremen",
    "hamburg",
    "berlin",
]

#### Bavaria

bavaria = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[0]}.pq"
)

bavaria["currentUse"] = None
bavaria["buildingNature"] = None
bavaria["heightAboveGround"] = None

#### Brandenburg

bb = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[1]}.pq"
)

bb["currentUse"] = bb["funktion"]
bb["buildingNature"] = bb["gebnutzbez"]
bb["heightAboveGround"] = None

#### BW

bw = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[2]}.pq"
)

bw["currentUse"] = None
bw["buildingNature"] = None
bw["heightAboveGround"] = None

#### Hessen

hessen = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[3]}.pq"
)

hessen["currentUse"] = hessen["funktion"]
hessen["buildingNature"] = hessen["gebnutzbez"]
hessen["heightAboveGround"] = None

#### MV

mv = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[4]}.pq"
)

mv["currentUse"] = mv["funktion"]
mv["buildingNature"] = mv["gebnutzbez"]
mv["heightAboveGround"] = None

#### NRW

nrw = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[5]}.pq"
)

nrw["currentUse"] = nrw["funktion"]
nrw["buildingNature"] = nrw["gebnutzbez"]
nrw["heightAboveGround"] = None

#### RLP

rlp = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[6]}.pq"
)

rlp["currentUse"] = rlp["funktion"]
rlp["buildingNature"] = rlp["gebnutzbez"]
rlp["heightAboveGround"] = None

#### Saarland

lots of random buildings, doesnt look like official data

saarland = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[7]}.pq"
)

saarland["currentUse"] = None
saarland["buildingNature"] = None
saarland["heightAboveGround"] = None

#### Sachsen

sachsen = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[8]}.pq"
)

sachsen["currentUse"] = sachsen["funktion"]
sachsen["buildingNature"] = sachsen["gebnutzbez"]
sachsen["heightAboveGround"] = None

#### sachsen-anhalt

sa = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_{ger_regions[9]}.pq"
)

sa["currentUse"] = sa["funktion"]
sa["buildingNature"] = sa["gebnutzbez"]
sa["heightAboveGround"] = None

#### schlesig holsten

sh = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[10]}.pq"
)

sh["currentUse"] = sh["funktion"]
sh["buildingNature"] = sh["gebnutzbez"]
sh["heightAboveGround"] = None

#### Thuringia

th = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[11]}.pq"
)

th["currentUse"] = th["funktion"]
th["buildingNature"] = th["gebnutzbez"]
th["heightAboveGround"] = None

#### niedersaxony

ni = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[12]}.pq"
)

ni["currentUse"] = None
ni["buildingNature"] = ni["gfk__bez"]
ni["heightAboveGround"] = None

#### bremen

bremen = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[13]}.pq"
).to_crs(epsg=3035)

bremen["currentUse"] = bremen["FUNKTION"]
bremen["buildingNature"] = bremen["GEBNUTZBEZ"]
bremen["heightAboveGround"] = None

#### hamburg

hamburg = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[14]}.pq"
).to_crs(epsg=3035)

hamburg["currentUse"] = None
hamburg["buildingNature"] = None
hamburg["heightAboveGround"] = None

#### berlin

berlin = gpd.read_parquet(
    f"/data/uscuni-ulce/buildings_standardised/buildings_germany_{ger_regions[15]}.pq"
)

berlin["currentUse"] = None
berlin["buildingNature"] = None
berlin["heightAboveGround"] = None

#### Merge Germany

print(
    ",".join(
        [
            "bavaria",
            "brandeburg",
            "bw",
            "hessen",
            "mv",
            "nrw",
            "rlp",
            "saarland",
            "sachsen",
            "sachsen-anhalt",
            "sh",
            "th",
            "ni",
            "bremen",
            "hamburg",
            "berlin",
        ]
    )
)

subset = ["currentUse", "buildingNature", "heightAboveGround", "geometry"]

all_gdfs = [
    bavaria,
    bb,
    bw,
    hessen,
    mv,
    nrw,
    rlp,
    saarland,
    sachsen,
    sa,
    sh,
    th,
    ni,
    bremen,
    hamburg,
    berlin,
]

%%time
all_germany = pd.concat((gdf[subset] for gdf in all_gdfs), ignore_index=True)

all_germany.to_parquet("/data/uscuni-ulce/buildings_standardised/merged_germany.pq")

### Austria

austria = gpd.read_parquet(DATADIR + regions[0]).to_crs(epsg=3035)

"""
'F_CODE', 'F_NAME', 'STRUCTURE FUNCTION', 'AGWR_OBJECT NUMBER', 'AGWR_TYPE',
'VERIFICATION_LB', 'DEL_FLAG', 'FLIGHT_YEAR', 'AS_DATE',
'RECORD_TYPE', 'GROUND_HEIGHT_MIN', 'RECORD_TYPE_GROUND_MIN',
'GROUND_HEIGHT_MEDIAN', 'RECORD_TYPE_GROUND_HEIGHT_MEDIAN',
'GROUND_HEIGHT_MAX', 'GROUND_TYPE_GROUND_MAX', 'OBJECT_HEIGHT_MEDIAN',
'OBJECT_TYPE_HEIGHT_MEDIAN', 'HEIGHT_OBJECT_MAX',
'RECORD_TYPE_HEIGHT_OBJECT_MAX', 'HEIGHT_OBJECT_EAVES',
'RECORD_TYPE_HEIGHT_OBJECT_EAVES', 'DATA SOURCE', 'DATA SOURCE_EXT_ID',
'CREATE_DATE', 'EDIT_DATE', 'OBJECTID', 'GLOBALID', 'SHAPE_AREA',
'geometry'
"""

austria["currentUse"] = austria["BAUWERKSFUNKTION"]
austria["buildingNature"] = austria["AGWR_TYP"]
austria["heightAboveGround"] = austria["HOEHE_OBJEKT_MEDIAN"]

### Slovakia

slovakia = gpd.read_parquet(DATADIR + regions[1]).to_crs(epsg=3035)

### Poland

poland = gpd.read_parquet(DATADIR + regions[2])

poland["currentUse"] = None
poland["buildingNature"] = None
poland["heightAboveGround"] = None

### Czechia

czechia = gpd.read_parquet(DATADIR + regions[3])

czechia["currentUse"] = None

### Merge all data

subset = ["currentUse", "buildingNature", "heightAboveGround", "geometry"]

all_gdfs = [all_germany, poland, austria, czechia, slovakia]

%%time
all_buildings = pd.concat((gdf[subset] for gdf in all_gdfs), ignore_index=True)

all_buildings.to_parquet(
    "/data/uscuni-ulce/buildings_standardised/merged_ce_buildings.pq"
)
