==> code/04_download_streets.ipynb <==
# Download street network

Download street network data from Overture Maps.

from core.generate_streets import read_overture_region_streets

import glob

import geopandas as gpd
import datetime

regions = "cadastre_regions_hull.parquet"

## Run sequentially for all streets

streets_dir = "/data/uscuni-ulce/overture_streets/"
regions_datadir = "/data/uscuni-ulce/"

region_hulls = gpd.read_parquet(
    regions_datadir + "regions/" + regions
)
region_hulls.shape

## Run in parallel 

Include a check for non-procesed streets in region_hulls and skip processed ones.

def process_and_save(region_hull, region_id):
    print("Processing", region_id, datetime.datetime.now())
    streets = read_overture_region_streets(region_hull, region_id)
    streets.to_parquet(streets_dir + f"streets_{region_id}.pq")

downloaded_streets = glob.glob(streets_dir + "*.pq")
downloaded_streets = [int(s.split("_")[-1][:-3]) for s in downloaded_streets]

region_hulls = region_hulls.loc[~region_hulls.index.isin(downloaded_streets),]
region_hulls

%%capture cap

from joblib import Parallel, delayed

n_jobs = -1
new = Parallel(n_jobs=n_jobs)(
    delayed(process_and_save)(region_hull.iloc[0], region_id)
    for region_id, region_hull in region_hulls.to_crs(epsg=4326).iterrows()
)
