==> code/04b_download_lithuania.ipynb <==
## Download buildings

import geopandas as gpd

buildings_url = 'https://inspire-geoportal.lt/resources/atom/bu/data/BU_INSPIRE2.zip'

!wget https://inspire-geoportal.lt/resources/atom/bu/data/BU_INSPIRE2.zip

!unzip BU_INSPIRE2.zip

buildings = gpd.read_file('BU_INSPIRE2.gml')

### Update region dict and save buildings

buildings = buildings.to_crs(epsg=3035)

# add to regions dir

bounds = buildings.bounds

from shapely import box

lithuania_convex_hull = box(bounds.minx.min(), bounds.miny.min(), bounds.maxx.max(), bounds.maxy.max())

regions_datadir = '/data/uscuni-ulce/'
region_hulls = gpd.read_parquet(
        regions_datadir + "regions/" + "cadastre_regions_hull.parquet"
    )

lithuania_region_id = region_hulls.index[-1] + 100

### add lithuania to the region_hulls
region_hulls.loc[lithuania_region_id, 'convex_hull'] = lithuania_convex_hull

region_hulls

region_hulls.to_parquet(regions_datadir + "regions/" + "cadastre_regions_hull.parquet")

buildings.to_parquet(f'/data/uscuni-ulce/buildings_standardised/buildings_lithuania.pq')

buildings.to_parquet(f'/data/uscuni-ulce/buildings_in_regions/buildings_{lithuania_region_id}.pq')

## Download Streets

import datetime
streets_dir = '/data/uscuni-ulce/overture_streets/'
from core.generate_streets import read_overture_region_streets
def process_and_save(region_hull, region_id):
    print('Processing', region_id, datetime.datetime.now())
    streets = read_overture_region_streets(region_hull, region_id)
    streets.to_parquet(streets_dir + f'streets_{region_id}.pq')

region_hull = region_hulls.to_crs(epsg=4326).loc[lithuania_region_id].iloc[0]
region_id = lithuania_region_id

%%time
process_and_save(region_hull, region_id)

gdf = gpd.read_parquet(streets_dir + f'streets_{lithuania_region_id}.pq')
