==> code/05_preprocess_cz_housing.ipynb <==
# Pre-process Czech housing estates

The data is extremely heterogenous. We use additional data to pre-process it prior morphological assessment.

import geopandas as gpd
import pandas as pd
import numpy as np

### Generate a GeoDataFrame that covers the socialist housing developments

!ls /data/uscuni-ulce/additional_data/housing_estates_locations

xcl = pd.read_excel(
    "/data/uscuni-ulce/additional_data/housing_estates_locations/HEs_delimitation.xlsx"
)
xcl.columns

estates = gpd.read_file(
    "/data/uscuni-ulce/additional_data/housing_estates_locations/ZSJ_2016.shp"
)
estates[
    [
        "KOD_ZSJ",
        "KOD_UTJ",
        "KOD_KU",
        "KOD_MOaMC",
        "KOD_OBEC",
        "KOD_ZUJ",
        "KOD_OKRES",
        "KOD_LAU1",
        "KOD_KRAJ",
        "KOD_CZNUTS",
    ]
]

# estates = gpd.read_file('/data/uscuni-ulce/additional_data/housing_estates_locations/ZSJ_2013.shp')
# estates[['KOD_ZSJ', 'KOD_UTJ', 'KOD_KU']]

estates["merge"] = estates["KOD_ZSJ"].astype(np.int64)

merged = pd.merge(
    estates,
    xcl[["BSU CODE6_2014", "SOC HE", "SOC&POSTSOC HE"]],
    left_on="merge",
    right_on="BSU CODE6_2014",
)

merged[merged["SOC HE"] == "yes"].shape

merged = merged[merged["SOC&POSTSOC HE"] == "yes"]

cz_soc_housing = merged.to_crs(epsg=3035)

### Process all buildings, in all regions that have socialist housing

from libpysal.graph import Graph

regions_datadir = "/data/uscuni-ulce/"
simplfied_buildings_dir = "/data/uscuni-ulce/processed_data/buildings/"
buildings_dir = "/data/uscuni-ulce/processed_data/buildings/"

region_hulls = gpd.read_parquet(
    regions_datadir + "regions/" + "cadastre_regions_hull.parquet"
)
inp, _ = cz_soc_housing.sindex.query(region_hulls.geometry, predicate="intersects")
czech_regions = region_hulls.iloc[np.unique(inp)]

%%time
for region_id, _ in czech_regions.iterrows():
    ## read region building polygons
    buildings = gpd.read_parquet(
        simplfied_buildings_dir + f"buildings_{region_id}.parquet"
    )
    buildings["area"] = buildings.area
    buildings["perimeter"] = buildings.length

    ### mark buildings in socialist housing
    _, blg_ix = buildings.sindex.query(cz_soc_housing.geometry, predicate="intersects")
    relevant = pd.Series(0, index=buildings.index, dtype=bool)
    relevant.iloc[np.unique(blg_ix)] = True

    # create a contiguity graph for the  buildings in socialist housing polygons
    contig = Graph.build_fuzzy_contiguity(buildings[relevant], buffer=1e-3)

    # get potetial socialist housing buildings
    min_area = 180
    min_perimeter = 55
    mask = (
        (buildings["area"] > min_area)
        & (buildings["perimeter"] > min_perimeter)
        & relevant
    )
    masked = buildings[mask].copy()

    ## merge potential socialist housing buildings, based on buffered graph contiguity
    masked.geometry = masked.buffer(1e-3)
    subgraph = contig.subgraph(masked.index)
    merged = masked.dissolve(subgraph.component_labels).explode()

    # add the rest of the data to the merged subset and save
    fixed = pd.concat(
        [buildings[~relevant], merged, buildings[(relevant) & (~mask)]],
        ignore_index=True,
    )
    assert not fixed.index.duplicated().any()
    fixed.drop(columns=["area", "perimeter"]).to_parquet(
        buildings_dir + f"buildings_{region_id}.parquet"
    )
