==> code/09_noise.ipynb <==
# Assign noise to clusters

import geopandas as gpd
import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier
from libpysal.graph import read_parquet
import numpy as np

v = "v10"

model_params = "_post_processing_v1"

clusters_dir = "/data/uscuni-ulce/processed_data/clusters/"
chars_dir = "/data/uscuni-ulce/processed_data/chars/"
regions_datadir = "/data/uscuni-ulce/"
region_hulls = gpd.read_parquet(
    regions_datadir + "regions/" + "cadastre_regions_hull.parquet"
)
graph_dir = "/data/uscuni-ulce/processed_data/neigh_graphs/"

cdata = pd.read_parquet(f"/data/uscuni-ulce/processed_data/clusters/cdata_{v}.pq")
# cdata = cdata.iloc[:, :-2]
cdata = cdata.drop("limLPS", axis=1)
cdata.index = cdata.index.astype(str)

def assign_noise(region_id, cdata):
    # read cluster data and characters
    clusters = pd.read_parquet(
        f"{clusters_dir}clusters_{region_id}_{v}.pq", columns=["final", "morph"]
    )
    chars = pd.read_parquet(f"{chars_dir}primary_chars_{region_id}.parquet")
    chars = chars[chars.index >= 0]
    assert (clusters.index == chars.index).all()

    # get morphotopes and noise groups values
    is_noise = clusters["final"].values == -1
    tq1 = read_parquet(graph_dir + f"tessellation_graph_{region_id}.parquet")
    noise_graph = tq1.subgraph(clusters[is_noise].index)
    components = noise_graph.component_labels
    noise_groups = chars[is_noise].groupby(components).median()

    # approax limAre
    noise_groups["limAre"] = (
        chars[is_noise]
        .groupby(components)["sdbAre"]
        .apply(lambda x: x.sort_values(ascending=True)[-min(30, x.shape[0]) :].sum())
    )

    noise_groups = noise_groups[cdata.columns]
    noise_groups.index = "-1_" + noise_groups.index.astype(str)

    all_data = pd.concat((cdata, noise_groups))
    scalar = StandardScaler()
    vals = scalar.fit_transform(all_data)
    all_data = pd.DataFrame(
        vals, index=all_data.index, columns=all_data.columns
    ).fillna(0)

    # setup tree data
    query_data = all_data.loc[noise_groups.index]
    tree_data = all_data.loc[cdata.index]

    ## query a knn classifier and assign predictions
    tree = KNeighborsClassifier(n_neighbors=5, n_jobs=-1)
    tree = tree.fit(X=tree_data.values, y=cdata.index.values)
    predictions = tree.predict(query_data.values)

    # map noise group clusters to building ids
    bid_to_ngroup = noise_graph.component_labels
    bid_to_prediction = bid_to_ngroup.map(
        pd.Series(
            predictions, query_data.index.str.split("_").str[-1].astype(int)
        ).to_dict()
    )
    clusters.loc[bid_to_prediction.index.values, "final"] = (
        bid_to_prediction.values.astype(int)
    )
    return clusters.final

%%time

for region_id, _ in region_hulls.iterrows():
    print(region_id)
    final_without_noise = assign_noise(region_id, cdata)
    clusters = gpd.read_parquet(f"{clusters_dir}clusters_{region_id}_{v}.pq")
    # assignment hasnt changed existing clusters
    pd.testing.assert_series_equal(
        final_without_noise[clusters.final != -1], clusters.final[clusters.final != -1]
    )
    clusters["final_without_noise"] = final_without_noise
    clusters.to_parquet(f'{clusters_dir}clusters_{region_id}_{v}.pq')
