==> code/12_postprocess_morphotope_ids.ipynb <==
# Post-process morphotope IDs to ensure IDs allocated to noise are unique per contiguity group

import geopandas as gpd

from libpysal import graph

v = "v10"

region_hulls = gpd.read_parquet(
    "/data/uscuni-ulce/regions/cadastre_regions_hull.parquet"
)

region_hulls

%%time

for region in region_hulls.index:
    df = gpd.read_parquet(
        f"/data/uscuni-ulce/processed_data/clusters/clusters_{region}_{v}.pq"
    )
    w = graph.read_parquet(
        f"/data/uscuni-ulce/processed_data/neigh_graphs/tessellation_graph_{region}.parquet"
    )
    noise = df[df.morph.str.endswith("-1")]
    noise_labels = w.subgraph(noise.index).component_labels
    df["unique_morph"] = df["morph"]
    df.loc[noise.index, "unique_morph"] = df.loc[noise.index, "unique_morph"].str[
        :-1
    ] + (noise_labels + 1).astype("str")
    df.to_parquet(f"/data/uscuni-ulce/processed_data/clusters/clusters_{region}_{v}.pq")

region = 69333

%%time

for region in region_hulls.index:
    buildings = gpd.read_parquet(
        f"/data/uscuni-ulce/processed_data/clusters/clusters_{region}_{v}.pq",
        columns=["morph", "unique_morph", "geometry"],
    )
    buildings_noise = buildings[buildings.morph.str.endswith("-1")]
    join_data = (
        buildings_noise.representative_point()
        .to_frame("geometry")
        .dissolve(buildings_noise.unique_morph)
        .representative_point()
    )

    mtopes = gpd.read_parquet(
        f"/data/uscuni-ulce/processed_data/morphotope_clusters/{v}/{region}_clusters.pq"
    ).reset_index()
    mtopes_noise = mtopes[mtopes.morph.str.endswith("-1")].copy()
    joined = mtopes_noise.reset_index().sjoin(join_data.to_frame(), how="left")
    cleaned = joined.groupby("index", as_index=False).first()
    mtopes["unique_morph"] = mtopes.morph.copy()
    mtopes.loc[cleaned["index"], "unique_morph"] = cleaned.unique_morph.values
    mtopes.to_parquet(
        f"/data/uscuni-ulce/processed_data/morphotope_clusters/{v}/{region}_clusters.pq"
    )
